convert_ast2cmdinvo.c 内の関数を別のファイルに移動
