+++ convert_ast2cmdinvo.c
@@ -1,115 +1,6 @@
 #include "minishell.h"
 
 /*
- * first に second を繋げて, firstだけfreeする
- */
-static char	*strjoin_and_free_first(char *first, char *second)
-{
-	char	*tmp;
-
-	tmp = first;
-	first = ft_strjoin(first, second);
-	free(tmp);
-	return (first);
-}
-
-/* string_nodeを文字列に戻す
- *うまくいけばASTの時点でこの文字列になるように依頼する
- */
-char	*string_node2string(t_parse_node_string *string_node)
-{
-	char	*result;
-
-	result = ft_strdup("");
-	while (string_node)
-	{
-		if (string_node->type == TOKTYPE_NON_EXPANDABLE)
-			result = strjoin_and_free_first(result, "\'");
-		else if (string_node->type == TOKTYPE_EXPANDABLE_QUOTED)
-			result = strjoin_and_free_first(result, "\"");
-		result = strjoin_and_free_first(result, string_node->text);
-		if (string_node->type == TOKTYPE_NON_EXPANDABLE)
-			result = strjoin_and_free_first(result, "\'");
-		else if (string_node->type == TOKTYPE_EXPANDABLE_QUOTED)
-			result = strjoin_and_free_first(result, "\"");
-		if (string_node->next)
-			string_node = string_node->next->content.string;
-		else
-			break ;
-	}
-	return (result);
-}
-
-/*
- * from から len文字substrして dst にくっつける
- */
-static char	*substr_and_join(char *dst, char *from, int len)
-{
-	char *tmp_dst;
-	char *tmp_substr;
-
-	tmp_dst = dst;
-	tmp_substr = ft_substr(from, 0, len);
-	if (dst)
-		dst = ft_strjoin(dst, tmp_substr);
-	else
-		dst = ft_strdup(tmp_substr);
-	free(tmp_dst);
-	free(tmp_substr);
-	return (dst);
-}
-
-/*
- * 環境変数展開した文字列を分解して返す
- */
-char	**split_expanded_str(char *str)
-{
-	char	**result;
-	char	quote_char;
-	int		len;
-	char	*text;
-
-	quote_char = '\0';
-	result = NULL;
-	while (*str)
-	{
-		// 空白飛ばし
-		while (*str && *str == ' ')
-			str++;
-		if (!*str)
-			break ;
-		len = 0;
-		text = NULL;
-		while (true)
-		{
-			// クオートの中じゃない時の空白は区切り
-			if ((!quote_char && str[len] == ' ') || !str[len])
-			{
-				text = substr_and_join(text, str, len);
-				str += len;
-				break;
-			}
-			if ((str[len] == '\'' || str[len] == '\"') && !(len > 0 && str[len - 1] == '\\'))
-			{
-				if (quote_char)
-					quote_char = '\0';
-				else
-					quote_char = str[len];
-				text = substr_and_join(text, str, len);
-				str += len + 1;  // クオートの次の文字
-				len = 0;
-			}
-			else
-				len++;
-		}
-		char	**tmp = result;
-		result = (char **)ptrarr_add_ptr((void **)tmp, text);
-		free(tmp);
-	}
-	return (result);
-}
-
-/*
  * 一度文字列を全て展開した上でexpanded_strarrを生成する
  */
 char	**expand_string_node(t_parse_node_string *string_node)
