+++ minishell.h
@@ -15,5 +15,6 @@ t_command_invocation	*cmd_ast_pipcmds2cmdinvo(t_parse_node_pipcmds *pipcmds);
 t_command_invocation	*cmd_ast_cmd2cmdinvo(t_parse_node_command *cmd_node);
 char					*string_node2string(t_parse_node_string *string_node);
 char					**expand_string_node(t_parse_node_string *string_node);
+char					**split_expanded_str(char *str);
 
 #endif
